Build under -Werror with Linux gcc and clang
tools/ci.sh only passed on macOS. On Ubuntu's gcc 13 the build stopped with
about 6,500 errors, and clang 18 failed too, before any test ran:

- The compiled world tables use designated initialisers on purpose, so a
  renamed field fails to compile (docs/design/pinatty.md). -Wpedantic calls
  them a C++20 extension. Silence that one note for C++ only: clang's
  -Wno-c++20-designator, gcc's -Wno-c++20-extensions (gcc has no finer flag).
  The language stays C++17.
- glm's headers are SYSTEM now, like imgui's, so gcc's -Wsign-conversion
  inside glm/detail/qualifier.hpp no longer lands on our targets.
- -Wnon-virtual-dtor applies to C++ only; gcc rejected it on the C sources.
- gcc's -Wconversion flags int-to-float where clang does not. Cast the loop
  counters and counts explicitly at each site; the values are unchanged.
- terrain_ops.h uses std::max({...}) and now includes <algorithm> itself
  instead of getting it transitively from libc++.
- Smaller gcc-only findings: two seed literals in synth.cpp (same values),
  a -Waddress false positive in respray_camera.h's static_assert, three
  misleading-indentation lines split (behaviour unchanged), a shadowed
  local in ragdoll_tests, a snprintf buffer in traffic_bench, and a default
  member initialiser on RoadSpine::deck_heights so designated test
  initialisers don't trip -Wmissing-field-initializers.

Gate on Ubuntu gcc 13: guards, configure and the -Werror build pass; 203 of
219 suites pass. The other 16 load the private cooked models under
assets/models/, which this checkout does not have.

+++ src/app/app.cpp
@@ -572,7 +572,7 @@ bool App::init() {
             for (int i = 0; i < 3; ++i) {
                 const float strength = 0.35f + 0.65f *
                     (0.5f + 0.5f * std::sin(phase - static_cast<float>(i)));
-                hud_.rect({left + i * 12.0f, y + 11}, {left + i * 12.0f + 5, y + 16},
+                hud_.rect({left + static_cast<float>(i) * 12.0f, y + 11}, {left + static_cast<float>(i) * 12.0f + 5, y + 16},
                           {0.86f, 0.20f, 0.14f, alpha * strength});
             }
             hud_.text(stage, {left + 52, y}, 19, {0.75f, 0.74f, 0.69f, alpha});
